refactor: InputView 분리
